cleanup

+++ src/rcpp_read_genome.cpp
@@ -73,7 +73,7 @@ Rcpp::List rcpp_read_genome (std::string fn,                                    
     
     char *sequence = faidx_fetch_seq(faidx, name, 0, length-1, &flen);          // try fetch sequence
     if (length!=flen) Rcpp::stop("Corrupted FASTA index. Delete and try again");// if fetched bytes differ from expected
-    for (size_t j=0; j<length; j++)
+    for (size_t j=0; j<(unsigned int)length; j++)
       sequence[j]=acgnt_filter_table[(unsigned char)sequence[j]];               // replace extended IUPAC with N
     
     rseq->emplace_back((const char*) sequence, length);                         // emplace sequence
